Feat: Created class Points and class CardReturn
Class points is a data class for storing the health, money, energy and mental variables,.
CardReturn is a data wrapper class, that consist of Points points and an String text, this class indicates what should be returned in the functions left and right from the class CardData.
CardData left and right functions changed from void to Function.